% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{arc}
\alias{arc}
\title{Arc between two nodes}
\usage{
arc(p0, p1, alpha = pi/3, n = 20L, radii = c(0, 0))
}
\arguments{
\item{p0, p1}{Numeric vector of length 2. Center coordinates}

\item{alpha}{Numeric scalar. Arc angle in radians.}

\item{n}{Integer scalar. Number of segments to approximate the arc.}

\item{radii}{Numeric vector of length 2. Radious}
}
\description{
Arc between two nodes
}
