% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.r
\name{arrow_fancy}
\alias{arrow_fancy}
\title{Arrow polygon.}
\usage{
arrow_fancy(x, alpha = 0, l = 0.25, a = pi/6, b = pi/1.5)
}
\arguments{
\item{x}{Numeric vector of length 2. Coordinates of the tip}

\item{alpha, l, a, b}{Numeric scalars}
}
\description{
Arrow polygon.
}
