% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart.R
\name{circle}
\alias{circle}
\title{Computes the coordinates to plot a circle}
\usage{
circle(x0, y0, r, rescale = TRUE)
}
\arguments{
\item{x0, y0}{Numeric scalars. Center of the circle}

\item{r}{Numeric scalar. Radius of the circle.}

\item{rescale}{Logical scalar. When \link{TRUE}, rescales the polygon
coordinates such that it preservers its aspect ratio once plotted.}
}
\description{
It actually does it by creating a polygon with 100 edges.
}
