% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{coords_calc_vertex}
\alias{coords_calc_vertex}
\title{Functions to calculate graph polygons coordinates}
\usage{
coords_calc_vertex(netenv, v)
}
\arguments{
\item{netenv}{An object of class network environment.}

\item{v, e}{Integer scalars. vertex or edge index.}
}
\description{
Functions to calculate graph polygons coordinates
}
