% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{edge_color_mixer}
\alias{edge_color_mixer}
\title{A wrapper of `rgb(colorRamp)`}
\usage{
edge_color_mixer(i, j, vcols, p = 0.5, alpha = 0.15)
}
\arguments{
\item{i, j}{Integer scalar. Indices of ego and alter from 1 through n.}

\item{vcols}{Vector of colors.}

\item{p}{Numeric scalar from 0 to 1. Proportion of mixing.}

\item{alpha}{Numeric scalar from 0 to 1. Passed to [grDevices:colorRamp]}
}
\value{
A color.
}
\description{
A wrapper of `rgb(colorRamp)`
}
