% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{fit_coords_to_dev}
\alias{fit_coords_to_dev}
\title{Adjust coordinates to fit aspect ratio of the device}
\usage{
fit_coords_to_dev(coords, adj = graphics::par("pin")[1:2])
}
\arguments{
\item{coords}{Two column numeric matrix. Vertices coordinates.}

\item{adj}{Numeric vector of length 2.}
}
\description{
Adjust coordinates to fit aspect ratio of the device
}
\details{
It first adjusts `coords` to range between `-1,1`, and then, using
`graphics::par("pin")`, it rescales the second column of it (`y`) to adjust
for the device's aspec ratio.
}
