% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{rescale_edge}
\alias{rescale_edge}
\title{Function to rescale the edge-width.}
\usage{
rescale_edge(width, rel = c(1, 3))
}
\arguments{
\item{width}{Numeric vector. width of the edges.}

\item{rel}{Vector of length 2 with the min and max width.}
}
\description{
Function to rescale the edge-width.
}
