% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{rescale_node}
\alias{rescale_node}
\title{Rescale the size of a node to make it relative to the aspect ratio of the device}
\usage{
rescale_node(size, rel = c(0.01, 0.05, 3))
}
\arguments{
\item{size}{Numeric vector. Size of the node (radious).}

\item{rel}{Numeric vector of length 3. Relative size for the minimum and maximum
of the plot, and curvature of the scale. The third number is used as `size^rel[3]`.}
}
\description{
Rescale the size of a node to make it relative to the aspect ratio of the device
}
\details{
This function is to be called after [plot.new], as it takes the parameter `usr`
from the
}
