% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{rescale_size}
\alias{rescale_size}
\title{Rescale the size of a node to make it relative to the aspect ratio of the device}
\usage{
rescale_size(size, rel = c(0.01, 0.05, 1))
}
\arguments{
\item{size}{Numeric vector. Size of the node (radious).}

\item{rel}{Numeric vector of length 3. Relative size for the minimum and maximum
of the plot, and curvature of the scale. The third number is used as \code{size^rel[3]}.}
}
\description{
Rescale the size of a node to make it relative to the aspect ratio of the device
}
\details{
This function is to be called after \link{plot.new}, as it takes the parameter \code{usr}
from the
}
