% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{rotate}
\alias{rotate}
\title{Rotation of polygon}
\usage{
rotate(mat, p0, alpha)
}
\arguments{
\item{mat}{Two-column numeric matrix. Coordinates of the polygon.}

\item{p0}{Numeric vector of length two. Origin.}

\item{alpha}{Numeric scalar. Rotation degree in radians.}
}
\description{
Rotation of polygon
}
