% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{set_colors}
\alias{set_colors}
\alias{set_colors.igraph}
\alias{set_colors.network}
\title{Create a vector of colors for the vertices}
\usage{
set_colors(x)

\method{set_colors}{igraph}(x)

\method{set_colors}{network}(x)
}
\arguments{
\item{x}{A graph.}
}
\description{
Create a vector of colors for the vertices
}
