% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_par.r
\name{update_par}
\alias{update_par}
\alias{update_edge_par}
\alias{update_vertex_par}
\title{Update graphical parameters of a \code{netplot} object}
\usage{
update_par(x, type, element, idx, ...)

update_edge_par(x, element, idx, ...)

update_vertex_par(x, element, idx, ...)
}
\arguments{
\item{x}{An object of class \code{netplot}.}

\item{type}{Character. Either \code{"edge"} or \code{"vertex"}.}

\item{element}{Character. If \code{"edge"}, then it can be either \code{"line"} or
\code{"arrow"}, otherwise it can be either \code{"core"} or \code{"frame"}.}

\item{idx}{(optional) Integer vector. Indices of the elements to be modified.
When missing, all elements are modified.}

\item{...}{Parameters to be modified. This is passed to \link[grid:editGrob]{grid::editGrob}. The
complete list of parameters that can be modified is available in \link[grid:gpar]{grid::gpar}.}
}
\value{
An object of class \code{netplot} with modified parameters.
}
\description{
\code{update_edge_par} and \code{update_vertex_par} are shorthands for
\code{update_par(type = "edge", ...)} and \code{update_par(type = "vertex", ...)}
respectively.
}
