% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot-package.R
\docType{package}
\name{netplot-package}
\alias{netplot-package}
\title{netplot: Beautiful Graph Drawing}
\description{
netplot is a graph visualization engine that emphasizes aesthetics while
providing default parameters that yield out-of-the-box nice visualizations.
It is built on top of the \strong{grid} graphics system (the same engine behind
\strong{ggplot2}) and works seamlessly with \code{igraph} and \code{network} objects.
}
\details{
The main entry point is \code{\link[=nplot]{nplot()}}. See \code{vignette("examples", package = "netplot")} for a general overview and \code{vignette("formulas", package = "netplot")} for mapping vertex/edge aesthetics from graph attributes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USCCANA/netplot}
  \item \url{https://usccana.github.io/netplot/}
  \item Report bugs at \url{https://github.com/USCCANA/netplot/issues}
}

}
\author{
\strong{Maintainer}: George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID})

Authors:
\itemize{
  \item George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID})
  \item Porter Bischoff \email{portergbischoff@gmail.com} (\href{https://orcid.org/0009-0004-6742-6281}{ORCID})
}

}
\keyword{internal}
\section{How to cite}{
  If you use \pkg{netplot} in published work, please cite it. Run
  \code{citation("netplot")} in R for the full entry.
}
