% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{nplot}
\alias{nplot}
\alias{netplot}
\alias{nplot.igraph}
\alias{nplot.network}
\alias{nplot.default}
\alias{print.netplot}
\title{Plot a network}
\usage{
nplot(...)

\method{nplot}{igraph}(x, layout = igraph::layout_nicely(x),
  vertex.size = igraph::degree(x, mode = "in"), vertex.col = "tomato",
  vertex.label = igraph::vertex_attr(x, "name"),
  edge.line.lwd = igraph::edge_attr(x, "weight"), ...)

\method{nplot}{network}(x, layout = sna::gplot.layout.kamadakawai(x, NULL),
  vertex.size = sna::degree(x, cmode = "indegree"), vertex.col = "tomato",
  vertex.label = network::get.vertex.attribute(x, "vertex.names"), ...)

\method{nplot}{default}(edgelist, layout, vertex.size = 1,
  bg.col = "transparent", vertex.nsides = 50, vertex.size.range = c(0.01,
  0.03), vertex.rot = 0, vertex.frame.prop = 0.2, vertex.label = NULL,
  vertex.label.range = c(5, 15), vertex.label.show = 1,
  edge.width.range = c(1, 2), edge.arrow.size = NULL,
  edge.curvature = pi/3, edge.line.breaks = 15, sample.edges = 1,
  skip.vertex = FALSE, skip.edges = FALSE, skip.arrows = skip.edges,
  add = FALSE, zero.margins = TRUE, gpar_order = c("vertex",
  "vertex.core", "vertex.frame", "vertex.label", "edge", "edge.line",
  "edge.arrow"), ...)

\method{print}{netplot}(x, y = NULL, newpage = TRUE, ...)
}
\arguments{
\item{...}{Further arguments passed to \code{\link[=set_gpar]{set_gpar()}}.}

\item{x}{An \code{igraph} object.}

\item{layout}{Numeric two-column matrix with the graph layout.}

\item{vertex.size}{Numeric vector of length \code{vcount(x)}. Absolute size of the vertex.}

\item{vertex.col, vertex.fill}{Color (border) and fill of the vertex (both core
and frame).}

\item{vertex.label}{Character vector of length \code{vcount(x)}. Labels.}

\item{edge.line.lwd}{Passed to \code{\link[=set_gpar]{set_gpar()}}.}

\item{edgelist}{An edgelist.}

\item{bg.col}{Color of the background.}

\item{vertex.nsides}{Numeric vector of length \code{vcount(x)}. Number of sizes of
the vertex. E.g. three is a triangle, and 100 approximates a circle.}

\item{vertex.size.range}{Vector of length \code{vcount(x)}.}

\item{vertex.rot}{Vector of length \code{vcount(x)}. Passed to \link[polygons:npolygon]{polygons::npolygon},
elevation degree from which the polygon is drawn.}

\item{vertex.frame.prop}{Vector of length \code{vcount(x)}. What proportion of the
vertex does the frame occupy (values between 0 and 1).}

\item{vertex.label.range}{Numeric vector of size 2 or 3. Relative scale of
\code{vertex.label.fontsize} in points (see \link[grid:gpar]{grid::gpar}).}

\item{vertex.label.show}{Numeric scalar. Proportion of labels to show as the
top ranking according to \code{vertex.size}.}

\item{edge.width.range}{Vector of length \code{ecount(x)}.}

\item{edge.arrow.size}{Vector of length \code{ecount(x)}.}

\item{edge.curvature}{Numeric vector of length \code{ecount(x)}. Curvature of edges
in terms of radians.}

\item{edge.line.breaks}{Vector of length \code{ecount(x)}. Number of vertices to
draw (approximate) the arc (edge).}

\item{sample.edges}{Numeric scalar between 0 and 1. Proportion of edges to sample.}

\item{skip.vertex, skip.edges, skip.arrows}{Logical scalar. When \code{TRUE} the object
is not plotted.}

\item{add}{Logical scalar.}

\item{zero.margins}{Logical scalar.}

\item{gpar_order}{Vector of character specifying in what other the graphical
parameters passed via \code{...} are evaluated.}

\item{y, ...}{Ignored}

\item{newpage}{Logical scalar. When \code{TRUE} calls \link[grid:grid.newpage]{grid::grid.newpage}.}
}
\value{
An object of class \code{c("netplot", "gTree", "grob", "gDesc")}. The object
has an additional set of attributes:
\itemize{
\item \code{.xlim, .ylim} vector of size two with the x-asis/y-axis limits.
\item \code{.layout} A numeric matrix of size \code{vcount(x) * 2} with the vertices positions
\item \code{.edgelist} A numeric matrix, The edgelist.
}
}
\description{
Plot a network
}
\details{
In the case of \code{edge.color}, the user can specify colors using \link{netplot-formulae}.
}
\examples{
library(igraph)
library(netplot)
set.seed(1)
x <- sample_smallworld(1, 200, 5, 0.03)

plot(x) # ala igraph
nplot(x) # ala netplot
}
