% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{nplot}
\alias{nplot}
\title{Plot a network}
\usage{
nplot(x, layout = igraph::layout_with_fr(x),
  vertex.size = igraph::degree(x), bg.col = "lightgray",
  vertex.shape = rep(100, igraph::vcount(x)), vertex.color = NULL,
  vertex.size.range = c(0.01, 0.03), vertex.frame.color = NULL,
  edge.width = NULL, edge.width.range = c(1, 2), edge.arrow.size = NULL)
}
\arguments{
\item{x}{An `igraph` object.}

\item{layout}{Numeric two-column matrix with the graph layout.}

\item{vertex.size}{Numeric vector of length `vcount(x)`. Absolute size of the vertex.}

\item{bg.col}{Color of the background.}

\item{vertex.shape}{Numeric vector of length `vcount(x)`. Number of sizes of
the vertex. E.g. three is a triangle, and 100 approximates a circle.}

\item{vertex.color}{Vector of length `vcount(x)`. Vertex colors.}

\item{vertex.size.range}{Vector of length `vcount(x)`.}

\item{vertex.frame.color}{Vector of length `vcount(x)`.}

\item{edge.width}{Vector of length `ecount(x)`.}

\item{edge.width.range}{Vector of length `ecount(x)`.}

\item{edge.arrow.size}{Vector of length `ecount(x)`.}
}
\description{
Plot a network
}
