% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.r
\name{nplot}
\alias{nplot}
\title{Plot a network}
\usage{
nplot(x, layout = igraph::layout_with_fr(x),
  vertex.size = igraph::degree(x), bg.col = "lightgray",
  vertex.shape = 100, vertex.color = NULL, vertex.size.range = c(0.01,
  0.03), vertex.frame.color = NULL, edge.width = NULL,
  edge.width.range = c(1, 2), edge.arrow.size = NULL,
  edge.color.mix = 0.5, edge.curvature = pi/3)
}
\arguments{
\item{x}{An \code{igraph} object.}

\item{layout}{Numeric two-column matrix with the graph layout.}

\item{vertex.size}{Numeric vector of length \code{vcount(x)}. Absolute size of the vertex.}

\item{bg.col}{Color of the background.}

\item{vertex.shape}{Numeric vector of length \code{vcount(x)}. Number of sizes of
the vertex. E.g. three is a triangle, and 100 approximates a circle.}

\item{vertex.color}{Vector of length \code{vcount(x)}. Vertex colors.}

\item{vertex.size.range}{Vector of length \code{vcount(x)}.}

\item{vertex.frame.color}{Vector of length \code{vcount(x)}.}

\item{edge.width}{Vector of length \code{ecount(x)}.}

\item{edge.width.range}{Vector of length \code{ecount(x)}.}

\item{edge.arrow.size}{Vector of length \code{ecount(x)}.}

\item{edge.color.mix}{Numeric vector of length \code{ecount(x)} with values in
\code{[0,1]}. 0 means color equal to ego's vertex color, one equals to alter's
vertex color.}

\item{edge.curvature}{Numeric vector of length \code{ecount(x)}. Curvature of edges
in terms of radians.}
}
\description{
Plot a network
}
\examples{
library(igraph)
library(netplot)
set.seed(1)
x <- sample_smallworld(1, 200, 5, 0.03)

plot(x) # ala igraph
nplot(x) # ala netplot
}
