% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_colors.r
\name{set_colors}
\alias{set_colors}
\alias{set_vertex_colors}
\title{Create a vector of colors for the vertices}
\usage{
set_colors(x, type, ...)

set_vertex_colors(x, vattr, is_categorical = FALSE,
  color_map = viridis::viridis)
}
\arguments{
\item{x}{A graph.}

\item{type}{Character. Either \code{edges} or \code{vertex}}

\item{vattr}{Character. Name of the vertex attribute.}

\item{is_categorical}{Logical. When \code{TRUE} sets the colors as categories.}

\item{color_map}{A function to generate a palette.}
}
\description{
Create a vector of colors for the vertices
}
