% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_mixer.r
\name{set_colors}
\alias{set_colors}
\alias{set_edges_colors}
\alias{set_vertex_colors}
\title{Create a vector of colors for the vertices}
\usage{
set_colors(x, type, ...)

set_edges_colors(x, vertex.color, ...)

set_vertex_colors(x, vattr, is_categorical = FALSE,
  color_map = viridis::viridis)
}
\arguments{
\item{x}{A graph.}

\item{type}{Character. Either \code{edges} or \code{vertex}}

\item{...}{Further arguments passed to the corresponding method.}

\item{vertex.color}{Vector of vertices colors.}

\item{vattr}{Name of the vertex attribute.}

\item{is_categorical}{Logical. When \code{TRUE} sets the colors as categories.}

\item{color_map}{A function to generate a palette.}
}
\description{
Create a vector of colors for the vertices
}
\details{
If no attribute is provided, then by defaul the colors are set accoring to
indegree.

\code{x} can be either a graph of class \code{igraph} or \code{network}.
}
\examples{
data(UKfaculty, package="igraphdata")
set_vertex_colors(UKfaculty, Group)
}
